+++ app/src/main/java/com/rocasspb/avaawaand/logic/TerrainUtils.kt
@@ -20,7 +20,7 @@ object TerrainUtils {
      */
     fun calculateTerrainMetrics(
         point: GeometryUtils.Point,
-        queryElevation: (GeometryUtils.Point) -> Double?
+        queryElevation: (GeometryUtils.Point) -> Int?
     ): TerrainMetrics? {
         val lng = point.x
         val lat = point.y
